Export FreeOpcUa Libraries
Currently, the header files for FreeOpcUa are being exported as part of the catkin package, but the library binaries are not also being exported.  If a downstream user attempts to include those header files, they will experience linking errors.  To address this, they have to dig the library target names out of the FreeOpcUa CMakeLists file and directly add them to target_link_libraries.

This commit addresses the issue by adding the four non-testing FreeOpCua libraries to the catkin_package() macro in the ros_opcua_impl_freeopcua CMakeLists.txt file.  This should allow users to freely include the already-exported headers without linking errors.